% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_model.R
\name{print.annotate_matrices}
\alias{print.annotate_matrices}
\title{Print Method for 'annotate_matrices'-Class Objects}
\usage{
\method{print}{annotate_matrices}(x, ...)
}
\arguments{
\item{x}{A 'annotate_matrices'-class object. The output of
\code{\link[=annotate_matrices]{annotate_matrices()}}.}

\item{...}{Optional arguments. To be passed to the default print method.}
}
\description{
Prints the output of \code{\link[=annotate_matrices]{annotate_matrices()}}.
}
\details{
It currently just print the list of matrices using
customized arguments for the default print method.
}
\examples{

# Adapted from https://lavaan.ugent.be/tutorial/cfa.html

library(lavaan)
HS.model <- '
visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9
'
fit_cfa <- cfa(HS.model,
               data = HolzingerSwineford1939)
annotate_matrices(fit_cfa)

}
