% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semunpack-package.R
\docType{package}
\name{semunpack-package}
\alias{semunpack-package}
\alias{_PACKAGE}
\title{semunpack: Functions for Learning and Teaching Structural Equation Modeling}
\description{
A collection of functions for learning and teaching structural equations, such as computing model df, matching paths to matrix elements, and plotting the estimation history.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfcheung.github.io/semunpack/}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

}
\keyword{internal}
